Handle simultaneous occurring extra attacks
If multiple SPELL_EXTRA_ATTACKS are generated from the same source,
e.g. an instant attack, it is not guaranteed that they will be
interleaved with their corresponding attack like this:

    SPELL_EXTRA_ATTACKS
    SWING_DAMAGE
    SPELL_EXTRA_ATTACKS
    SWING_DAMAGE

This sequence is possible:

    SPELL_EXTRA_ATTACKS
    SPELL_EXTRA_ATTACKS
    SWING_DAMAGE
    SWING_DAMAGE

Keep track of how many extra attacks are generated in the "moment"
(decided by the allowed timestamp discrepancy in swingTimerCLEUCheck)
and skip the correct number of swings.