Let swing timer not reset on extra special attacks
If a heroic strike or cleave is queued before an extra attack is
triggered, the extra attack will be a cast heroic strike or cleave
instead of a normal attack.

Previously a special extra attack was not checked for so the swing
timer would reset erroneously.

Handle spells that modify the auto attack in swingTimerCLEUCheck now
to allow extra attack detection.